int to double triangleArea

+++ proj.java
@@ -7,8 +7,9 @@ public class proj {
 		double area= a*b;
 		System.out.println(area);
 		//Update this project
-				int triangleArea=1/2*a*b;
-				System.out.println(triangleArea);
+		double triangleArea=1/2*a*b;
+		System.out.println(triangleArea);
+		
 	}
 
 }
